moved all nytimes list into a seperate file and changed routes
